fix date picking in Add item test.
Move date picking util to TestHelpers
