+++ app/src/androidTest/java/com/example/househomey/testUtils/TestHelpers.java
@@ -2,11 +2,15 @@ package com.example.househomey.testUtils;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.clearText;
+import static androidx.test.espresso.action.ViewActions.click;
 import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
 import static androidx.test.espresso.action.ViewActions.pressImeActionButton;
+import static androidx.test.espresso.action.ViewActions.replaceText;
 import static androidx.test.espresso.action.ViewActions.scrollTo;
 import static androidx.test.espresso.action.ViewActions.typeText;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static androidx.test.espresso.matcher.ViewMatchers.withContentDescription;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 
@@ -14,10 +18,18 @@ import android.app.Activity;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.net.Uri;
+import android.view.View;
+import android.view.ViewGroup;
+import android.view.ViewParent;
 
 import androidx.annotation.IdRes;
 
 import com.example.househomey.R;
+
+import org.hamcrest.Description;
+import org.hamcrest.Matcher;
+import org.hamcrest.Matchers;
+import org.hamcrest.TypeSafeMatcher;
 
 /*
  * Helper methods for Espresso testing in Android applications.
@@ -96,5 +108,47 @@ public class TestHelpers {
      */
     public static Uri mockImageUri(int resourceId) {
         return Uri.parse("android.resource://com.example.househomey/" + resourceId);
+    }
+
+    /*
+     * Sets the Material DatePicker date for a specified date text input.
+     * Uses text input mode of the DatePicker.
+     *
+     * @param id   The resource ID of the view element for which the date is being set.
+     * @param date The date string to be set.
+     */
+    public static void pickDate(int id, String date) {
+        onView(withId(id)).perform(click());
+        onView(withContentDescription("Switch to text input mode")).perform(click());
+        onView(Matchers.allOf(childAtPosition(childAtPosition(
+                        withId(com.google.android.material.R.id.mtrl_picker_text_input_date), 0), 0),
+                isDisplayed())).perform(replaceText(date), closeSoftKeyboard());
+        onView(withText("OK")).perform(click());
+    }
+
+    /*
+     * Custom Matcher for locating a child at a specific position within a parent
+     *
+     * @param parentMatcher The Matcher<View> used to match the parent View or ViewGroup
+     * @param position      The position of the child View within the parent.
+     * @return A Matcher<View> instance for finding a child View at the specified position.
+     */
+    public static Matcher<View> childAtPosition(
+            final Matcher<View> parentMatcher, final int position) {
+
+        return new TypeSafeMatcher<View>() {
+            @Override
+            public void describeTo(Description description) {
+                description.appendText("Child at position " + position + " in parent ");
+                parentMatcher.describeTo(description);
+            }
+
+            @Override
+            public boolean matchesSafely(View view) {
+                ViewParent parent = view.getParent();
+                return parent instanceof ViewGroup && parentMatcher.matches(parent)
+                        && view.equals(((ViewGroup) parent).getChildAt(position));
+            }
+        };
     }
 }
